libflo: A C++ library for parsing Flo files
Chisel has a "flo" backend, which outputs "flo" files.  These are
essentially just representations of the internal Chisel graph format,
which probably makes it the easiest way to interface external tools
with Chisel.

This code originated while trying to write a DREAMER ISA simulator but
appears to be useful outside of just that so it's been split off into
a library.