Import Flo parser code from DREAMER
This commit contains the Flo file parser that I'd written while trying
to figure out how to write a DREAMER toolchain over the last few
months.  This has gone through a number of iterations on the source so
it's not just a straight-up copy so I'm not super confident in the
code...

There's also a really simple test case that just ensures that this can
parse _something_ and output it.