Copy operation list from Assembler.cpp
This appears to be the proper definition of what operations a Flo file
can contain.  It's technically the list of DREAMER instructions, but
they're still very coupled so I think it should be fine.

Note that a lot of these aren't implemented because I'm not entirely
sure what they do...