Hack around CAT width reporting
Essentially I've got a special method in here (node::with_cat_width)
that sets the output width of a CAT operation.  Unfortunately, you
can't know this from just looking at a single line of the Flo file
(which is kind of a pain).

This makes outwidth() report correctly for a CAT node, but note that
width() is silently wrong for one of the nodes.  If we have

  d = cat s0 s1

then outwid() reports the width of d and width() reports the width of
s1.  You'll have to manually determine the width of s0 (via
subtraction) yourself.