Abort width inference early when possible
While the full width inference algorithm is O(n^2) in the worst case,
that's really uncommon.  Most of the time it will just bail out after
a single loop because there's no work to do at all.

This adds an early out to avoid the worst-case behavior in broken
graphs.