Add RSHD opcode
The general idea here is that I want to make the multi-word
expansion's input nodes work.  This "RSHD" opcode is designed to work
in pretty much exactly the same way a "CATD" works: it's only usable
in debug mode and is used to encode the name mangling within the Flo
file.