Infer the width of constant RSH offsets
The general idea here is that RSH nodes frequently have a constant
attached to them, and the exact width of those nodes doesn't really
matter because nothing will actually end up using them.  I just use
Chisel's rule of picking the smallest bit width that will fit the
constant (with a special case for 0) in order to match things.