Don't list memory nodes twice
As part of the pflo parsing code I had to build a map that had both
versions of a node's name.  This unfortunately ended up resulting in
having memory nodes added twice under some circumstances, which is
wrong.

I "fixed" this by building two maps: one with only the node name and
another with both the node name and the placed bits.