Allow users to use non-strict width inference
The general idea here is that some tools produce malformed Flo files
and we'd just like to pass them through to see what happens.  This
patch adds a mode that parses Flo files in just that mode.  Note that
this fixes the graph as it reads the file, which means it'll produce a
sanitized Flo file in memory -- that Flo file just might not actually
consist of anything sane!

In order to run in this mode, do something like

  $ LIBFLO__ALLOW_BAD_WIDTHS="true" flo-canon ...

which will enable the new mode.