Use a text position instead of (x,y)
The general idea here is that I want to avoid having libflo depend on
a mesh network.  This allows me to actually use this to parse placed
non-mesh networks, which is necessary for when I split up the
placement and routing tools.

Unfortunately this is yet another backwards-incompatible change, but
it's really just fixing that last one in a sane manner.