Allow users to specify a node creation function
The general idea here is that some subclasses of libflo::node will
need to have extra information passed to them in order to be
construction.  An example is in dreamer-par, where the scheduler needs
to have a pointer to a class that describes the tile that owns each
node (as opposed to just the name).

What this does is add a function to flo::parse() that sits between the
parser and creating a node, thus allowing users to do anything right
there.