Strip invalid nodes from the node set
I have some code (flo-mwe, for example) that walks the list of Flo
nodes and does something with them.  It expects these nodes to be in a
valid state (all widths inferred), but it turns out that because of a
Chisel bug sometimes there are just little disconnected fragments of
the circuit output that can't have anything sane done with them.

I used to just strip the operations, but this strips the nodes as
well.

This should have been part of "ca88752 Work around a Chisel bug".