add a class_name method
When generating code (be it C++, LLVM, or Verilog) from a flo file,
it is useful to know what the top-level class name of the flo file is.
The flo-llvm code already includes a method like this in its
parameterized subclass of libflo's flo class. This commit adds the
method to libflo's flo class in order to reduce code duplication.

Testing done: Deleted the existing "class_name" method in flo-llvm to
force it to use the new method and ran the tests. All tests passed
successfully.