Be a bit stricter about RD's first operand
When Howard was trying to figure out what exactly some Flo nodes do,
he asked me something like "What is the first operand for RD supposed
to do?".  I think it's a read-enable, but as I'm under the impression
you can't trigger those from Chisel I'm not exactly sure.

This patch just makes sure that whatever is passed in is exactly the
constant "1", as that's all I've ever seen.  I figure it's best to be
too strict than too lose...