Update the hand-written tests to the latest Chisel
So it appears that Chisel always outputs a "1" in every REG's S
operand.  I don't know or care when that changed, but it appears that
at some point in the past that wasn't true.  As the old operand seems
to be useless, I'm just removing it from all of these tests.