Reg->o(1) is an enable, not an init
I wasn't sure exactly what this argument to reg nodes was, and it
turns out my guess was wrong!  This means that all the width infrence
I was doing was wrong, because enable is a 1-bit signal (and init
would have been an N-bit signal).

This patch:
* Changes the width inference to infer the correct width.
* Eliminates the check to make sure o(1) is a constant, and of the
  value 1.
* Modifies all the test cases to check for the correct width inference
  rules.
* Adds a test that contains a regisetr enable signal.

This all had to be done at once.